+++ src/LAPACKReplacement.cpp
@@ -120,29 +120,32 @@ extern "C" void spotrf_(	const char *uplo, const unsigned long *_n, float *a,
 
 extern "C" void dtrtrs_(	const char *UPLO, const char *TRANS, const char *DIAG,
 							const unsigned long *N, const unsigned long *NRHS,
-							double *A, const unsigned long *LDA, double *B, const unsigned long *LDB, int *INFO )
+							double *A, const unsigned long *LDA, double *B, const unsigned long *LDB, long *INFO
+							)
 {
 	; /* Dummy. If SQProblemSchur is to be used, system LAPACK must be used */
 }
 
 extern "C" void strtrs_(	const char *UPLO, const char *TRANS, const char *DIAG,
 							const unsigned long *N, const unsigned long *NRHS,
-							float *A, const unsigned long *LDA, float *B, const unsigned long *LDB, int *INFO )
+							float *A, const unsigned long *LDA, float *B, const unsigned long *LDB, long *INFO
+							)
 {
 	; /* Dummy. If SQProblemSchur is to be used, system LAPACK must be used */
 }
 
 extern "C" void dtrcon_(	const char *NORM, const char *UPLO, const char *DIAG,
 							const unsigned long *N, double *A, const unsigned long *LDA,
-							double *RCOND, double *WORK, const unsigned long *IWORK, int *INFO )
+							double *RCOND, double *WORK, const unsigned long *IWORK, long *INFO
+							)
 {
 	; /* Dummy. If SQProblemSchur is to be used, system LAPACK must be used */
 }
 
 extern "C" void strcon_(	const char *NORM, const char *UPLO, const char *DIAG,
 							const unsigned long *N, float *A, const unsigned long *LDA,
-							float *RCOND, float *WORK, const unsigned long *IWORK, int *INFO )
+							float *RCOND, float *WORK, const unsigned long *IWORK, long *INFO
+							)
 {
 	; /* Dummy. If SQProblemSchur is to be used, system LAPACK must be used */
 }
-
